Completed docs for display, audio, and buttons
