+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "sprig-essentials"
-version = "0.2.0"
+version = "0.3.0"
 authors = [
   { name="Sooraj Sannabhadti", email="developer.soorajs@gmail.com" },
 ]
