+++ src/sprig-essentials/display.py
@@ -60,10 +60,10 @@ class Display:
     # Turn on backlight as it doesn't turn on automatically
     def startBacklight(self,
                        backlight_pin: pin_number) -> digitalio__digital_in_out:
-        led = digitalio.DigitalInOut(backlight_pin)
-        led.direction = digitalio.Direction.OUTPUT
-        led.value = True
-        return led
+        backlight = digitalio.DigitalInOut(backlight_pin)
+        backlight.direction = digitalio.Direction.OUTPUT
+        backlight.value = True
+        return backlight
 
     def createSPI(self,
                   clock: pin_number,
@@ -103,16 +103,10 @@ class Display:
         group.append(sprite)
 
 
-# Convert rgb colour values to hex
-def convertRGBToHex(rgb: list):
-    if len(rgb) > 3:
-        raise IndexError(f"The list should have 3 value for the red, green, and blue channels! You have: {len(rgb)} values in your list.")
-    elif rgb[0] > 255 or rgb[0] < 0 or rgb[1] > 255 or rgb[1] < 0 or rgb[2] > 255 or rgb[2] < 0:
-        raise ValueError(f"The values should be between 0 and 255 inclusive! Your input is: {rgb}.")
-    elif (type(rgb[0]) != int and type(rgb[0]) != float) or (type(rgb[1]) != int and type(rgb[1]) != float) or (type(rgb[2]) != int and type(rgb[2]) != float):
-        raise TypeError("The values in the list are not of the correct type! It should either be an int or a float.")
-    else:
-        return int("{:02x}{:02x}{:02x}".format(rgb[0], rgb[1], rgb[2]), 16)
+# Releases displays to create new ones if needed
+def releaseDisplays():
+    # Reset all pins to allow new connections
+    displayio.release_displays()
 
 def createDisplayGroup(x: int = 0,
                        y: int = 0,
@@ -145,10 +139,10 @@ def createSprite(bitmap: displayio__bitmap,
     return sprite
 
 def createTextSprite(text: str,
-                     colour: list,
+                     colour: tuple,
                      x: int = 0,
                      y: int = 0,
                      font: terminalio__font = terminalio.FONT) -> label__label:
-    text_area = label.Label(font, text=text, color=colour[-1], x=x, y=y)
+    text_area = label.Label(font, text=text, color=colour, x=x, y=y)
 
     return text_area
