+++ src/sprig-essentials/graphics/__init__.py
@@ -1,0 +1,1 @@
+# Useful functions to have when using the graphics module
