+++ src/sprig-essentials/graphics/shapes.py
@@ -1,0 +1,5 @@
+import board
+
+class Shape:
+    def __init__(self) -> None:
+        pass
