+++ src/sprig-essentials/graphics/text.py
@@ -1,0 +1,5 @@
+import board
+
+class Text:
+    def __init__(self) -> None:
+        pass
